+++ backend/libraries/chat_events/src/lib.rs
@@ -9,6 +9,7 @@ mod expiring_events;
 mod hybrid_map;
 mod last_updated_timestamps;
 mod message_content_internal;
+mod message_ids;
 mod metrics;
 mod search_index;
 mod stable_memory;